feat(browser-capture): phase 1 - decompose monolith
