+++ packages/browser-capture/__tests__/test-generator.test.js
@@ -1,0 +1,60 @@
+import { describe, expect, it } from 'vitest';
+import { generatePlaywrightTest } from '../src/interactions/test-generator.js';
+
+describe('generatePlaywrightTest', () => {
+  it('produces a Playwright test with expected structure', () => {
+    const url = 'http://localhost:5173/dashboard';
+    const output = generatePlaywrightTest(url, [
+      {
+        type: 'click',
+        element: { testId: 'refresh' },
+        react: { componentName: 'DashboardHeader' },
+      },
+      {
+        type: 'fill',
+        element: { testId: 'search' },
+        value: 'widgets',
+      },
+      {
+        type: 'navigate',
+        url: 'http://localhost:5173/settings',
+      },
+    ]);
+
+    expect(output).toContain(
+      "import { test, expect } from '@playwright/test';",
+    );
+    expect(output).toContain(
+      '// Generated by copilot-devtools record-interactions',
+    );
+    expect(output).toContain(`// Source: ${url}`);
+    expect(output).toContain(
+      'test("recorded: localhost/dashboard", async ({ page }) => {',
+    );
+    expect(output).toContain(
+      'await page.goto("http://localhost:5173/dashboard");',
+    );
+    expect(output).toContain('await page.getByTestId("refresh").click();');
+    expect(output).toContain(' // DashboardHeader');
+    expect(output).toContain(
+      'await page.getByTestId("search").fill("widgets");',
+    );
+    expect(output).toContain(
+      'await page.goto("http://localhost:5173/settings");',
+    );
+    expect(output.endsWith('\n')).toBe(true);
+  });
+
+  it('JSON-serializes captured values to prevent injection', () => {
+    const output = generatePlaywrightTest('http://localhost:5173', [
+      {
+        type: 'fill',
+        element: { testId: 'comment' },
+        value: '"); await evil(); //',
+      },
+    ]);
+
+    expect(output).toContain('.fill("\\"); await evil(); //")');
+    expect(output).not.toContain('.fill(""); await evil(); //")');
+  });
+});
