+++ packages/browser-capture/package.json
@@ -6,10 +6,17 @@
   "bin": {
     "copilot-devtools": "./bin/copilot-devtools.js"
   },
+  "scripts": {
+    "test": "vitest run",
+    "test:watch": "vitest"
+  },
   "dependencies": {
     "@jridgewell/trace-mapping": "0.3.31",
     "@modelcontextprotocol/sdk": "1.29.0",
     "playwright-core": "catalog:",
     "zod": "3.25.76"
+  },
+  "devDependencies": {
+    "vitest": "catalog:"
   }
 }
