+++ packages/browser-capture/src/capture/interactions.js
@@ -1,0 +1,136 @@
+import fs from 'node:fs';
+import path from 'node:path';
+import { buildMetadata } from '../artifacts/metadata.js';
+import { ensureArtifactsDirectory } from '../artifacts/paths.js';
+import { writeJson } from '../artifacts/write.js';
+import { requireUrl, resolveDurationMs } from '../cli/args.js';
+import { connectCDP } from '../cdp/connect.js';
+import { httpGetJson } from '../cdp/http.js';
+import { log } from '../config/log.js';
+import { isSanitizeEnabled } from '../config/runtime.js';
+import { interactionRecorderInject } from '../inject/paths.js';
+import { resolveSourceLocation } from '../interactions/source-map.js';
+import { generatePlaywrightTest } from '../interactions/test-generator.js';
+import { sanitizeArtifacts } from '../sanitize/index.js';
+
+export async function recordInteractions(url, options = {}) {
+  const targetUrl = requireUrl('record-interactions', url);
+  const durationMs = resolveDurationMs(options);
+  const artifactsDir = ensureArtifactsDirectory('interactions');
+  const browserInfo = await httpGetJson('/json/version');
+
+  const browser = await connectCDP();
+  const context = await browser.newContext();
+  const interactions = [];
+  let captureResult = null;
+
+  try {
+    const page = await context.newPage();
+
+    await page.exposeFunction('__recordInteraction', (event) => {
+      interactions.push({ ...event, timestamp: Date.now() });
+    });
+    await page.addInitScript({ path: interactionRecorderInject });
+
+    await page.goto(targetUrl, { waitUntil: 'load', timeout: 30_000 });
+    log(`Recording interactions on ${targetUrl} for ${durationMs / 1000}s...`);
+    log('Interact with the page. Interactions are captured automatically.');
+
+    await page.waitForTimeout(durationMs).catch(() => {});
+
+    const activeInput = await page
+      .evaluate(() => {
+        const el = document.activeElement;
+        if (!el) return null;
+        const tag = el.tagName?.toLowerCase();
+        if (
+          (tag === 'input' || tag === 'textarea') &&
+          el.type !== 'checkbox' &&
+          el.type !== 'radio' &&
+          el.type !== 'file'
+        ) {
+          return {
+            tag,
+            value: el.value,
+            id: el.id || null,
+            testId: el.getAttribute('data-testid') || null,
+            name: el.getAttribute('name') || null,
+          };
+        }
+        return null;
+      })
+      .catch(() => null);
+
+    if (activeInput?.value) {
+      const last = interactions.at(-1);
+      const isDuplicate =
+        last?.type === 'fill' && last?.value === activeInput.value;
+      if (!isDuplicate) {
+        interactions.push({
+          type: 'fill',
+          element: {
+            tag: activeInput.tag,
+            id: activeInput.id,
+            testId: activeInput.testId,
+            name: activeInput.name,
+          },
+          value: activeInput.value,
+          react: null,
+          timestamp: Date.now(),
+        });
+      }
+    }
+
+    for (const interaction of interactions) {
+      if (interaction.react?.scriptUrl && !interaction.react?.fileName) {
+        const resolved = await resolveSourceLocation(
+          interaction.react.scriptUrl,
+          interaction.react.lineNumber ?? 1,
+          interaction.react.columnNumber ?? 0,
+        );
+        if (resolved) interaction.react.originalSource = resolved;
+      }
+    }
+
+    const pageDetails = await page
+      .evaluate(() => ({ url: location.href, title: document.title }))
+      .catch(() => ({ url: null, title: null }));
+
+    const metadata = buildMetadata('interactions', artifactsDir, browserInfo, {
+      url: pageDetails.url || targetUrl,
+      title: pageDetails.title || null,
+      durationMs,
+      interactionCount: interactions.length,
+    });
+
+    writeJson(artifactsDir, 'metadata.json', metadata);
+    writeJson(artifactsDir, 'interactions.json', { interactions });
+
+    if (isSanitizeEnabled()) sanitizeArtifacts(artifactsDir);
+
+    const sanitizedInteractions = isSanitizeEnabled()
+      ? JSON.parse(
+          fs.readFileSync(path.join(artifactsDir, 'interactions.json'), 'utf8'),
+        ).interactions
+      : interactions;
+    const testCode = generatePlaywrightTest(targetUrl, sanitizedInteractions);
+    const testFilePath = path.join(artifactsDir, 'generated.test.ts');
+    fs.writeFileSync(testFilePath, testCode, 'utf8');
+
+    log(`Captured ${interactions.length} interaction(s).`);
+    log(`Generated test: ${testFilePath}`);
+    log(`Saved interaction artifacts to ${artifactsDir}`);
+
+    captureResult = {
+      artifactsDir,
+      metadata,
+      interactionCount: interactions.length,
+      testFile: testFilePath,
+    };
+  } finally {
+    await context.close().catch(() => {});
+    await browser.close().catch(() => {});
+  }
+
+  return captureResult;
+}
