+++ packages/browser-capture/src/cli/args.js
@@ -1,0 +1,71 @@
+import { getDefaultDurationMs } from '../config/env.js';
+
+export function parseArgs(argv) {
+  const positionals = [];
+  const options = {};
+
+  for (let index = 0; index < argv.length; index += 1) {
+    const arg = argv[index];
+
+    if (!arg.startsWith('--')) {
+      positionals.push(arg);
+      continue;
+    }
+
+    const withoutPrefix = arg.slice(2);
+    const [key, inlineValue] = withoutPrefix.split('=', 2);
+
+    if (inlineValue !== undefined) {
+      options[key] = inlineValue;
+      continue;
+    }
+
+    const next = argv[index + 1];
+    if (next && !next.startsWith('--')) {
+      options[key] = next;
+      index += 1;
+    } else {
+      options[key] = true;
+    }
+  }
+
+  return { positionals, options };
+}
+
+export function requireUrl(command, url) {
+  if (!url) {
+    throw new Error(`${command} requires a URL.`);
+  }
+
+  return url;
+}
+
+export function resolveDurationMs(options) {
+  if (options['duration-ms']) {
+    const durationMs = Number(options['duration-ms']);
+    if (!Number.isFinite(durationMs) || durationMs <= 0) {
+      throw new Error('duration-ms must be a positive number.');
+    }
+
+    return durationMs;
+  }
+
+  if (options.duration) {
+    const durationSeconds = Number(options.duration);
+    if (!Number.isFinite(durationSeconds) || durationSeconds <= 0) {
+      throw new Error('duration must be a positive number of seconds.');
+    }
+
+    return Math.round(durationSeconds * 1000);
+  }
+
+  return getDefaultDurationMs();
+}
+
+export function captureOptions(options) {
+  return {
+    durationMs: resolveDurationMs(options),
+    attach: options.attach === true,
+    sanitize: options['no-sanitize'] !== true,
+  };
+}
