+++ packages/browser-capture/src/config/log.js
@@ -1,0 +1,6 @@
+import { isMcpServer } from './runtime.js';
+
+export function log(...args) {
+  if (isMcpServer()) process.stderr.write(args.join(' ') + '\n');
+  else console.log(...args);
+}
