+++ packages/browser-capture/src/inject/performance-observer.inject.js
@@ -1,0 +1,86 @@
+(() => {
+  const state = {
+    navigationStart: performance.timeOrigin,
+    lcpEntries: [],
+    clsEntries: [],
+    clsValue: 0,
+    inpEntries: [],
+    observerErrors: [],
+  };
+
+  const safeObserve = (type, callback, options) => {
+    try {
+      const observer = new PerformanceObserver((list) => {
+        try {
+          callback(list.getEntries());
+        } catch (error) {
+          state.observerErrors.push({
+            type,
+            message: error instanceof Error ? error.message : String(error),
+          });
+        }
+      });
+      observer.observe(options);
+    } catch (error) {
+      state.observerErrors.push({
+        type,
+        message: error instanceof Error ? error.message : String(error),
+      });
+    }
+  };
+
+  safeObserve(
+    'largest-contentful-paint',
+    (entries) => {
+      state.lcpEntries.push(
+        ...entries.map((entry) => ({
+          entryType: entry.entryType,
+          startTime: entry.startTime,
+          duration: entry.duration,
+          size: entry.size || null,
+          url: entry.url || null,
+          id: entry.id || null,
+          renderTime: entry.renderTime || null,
+          loadTime: entry.loadTime || null,
+        })),
+      );
+    },
+    { type: 'largest-contentful-paint', buffered: true },
+  );
+
+  safeObserve(
+    'layout-shift',
+    (entries) => {
+      for (const entry of entries) {
+        if (!entry.hadRecentInput) {
+          state.clsValue += entry.value;
+        }
+        state.clsEntries.push({
+          entryType: entry.entryType,
+          startTime: entry.startTime,
+          value: entry.value,
+          hadRecentInput: entry.hadRecentInput,
+        });
+      }
+    },
+    { type: 'layout-shift', buffered: true },
+  );
+
+  safeObserve(
+    'event',
+    (entries) => {
+      state.inpEntries.push(
+        ...entries.map((entry) => ({
+          entryType: entry.entryType,
+          name: entry.name,
+          startTime: entry.startTime,
+          duration: entry.duration,
+          interactionId: entry.interactionId || 0,
+        })),
+      );
+    },
+    { type: 'event', buffered: true, durationThreshold: 0 },
+  );
+
+  window.__COPILOT_DEVTOOLS_PERF__ = state;
+})();
