+++ packages/browser-capture/src/interactions/test-generator.js
@@ -1,0 +1,78 @@
+import { interactionToLocator } from './locator.js';
+
+export function generatePlaywrightTest(url, interactions) {
+  const actionLines = [];
+
+  for (const interaction of interactions) {
+    const locator = interactionToLocator(interaction);
+    let comment = '';
+    const r = interaction.react;
+    if (r) {
+      const src = r.originalSource
+        ? `${r.originalSource.source}:${r.originalSource.line}`
+        : r.fileName
+          ? `${r.fileName}:${r.lineNumber}`
+          : null;
+      const name = r.componentName
+        ? r.componentName.replace(/\*\//g, '')
+        : null;
+      if (name || src) comment = ` // ${[name, src].filter(Boolean).join(' ')}`;
+    }
+
+    switch (interaction.type) {
+      case 'navigate':
+        actionLines.push(
+          `  await page.goto(${JSON.stringify(interaction.url)});`,
+        );
+        break;
+      case 'click':
+        if (!locator) break;
+        actionLines.push(`  await ${locator}.click();${comment}`);
+        break;
+      case 'fill':
+        if (!locator) break;
+        actionLines.push(
+          `  await ${locator}.fill(${JSON.stringify(interaction.value ?? '')});${comment}`,
+        );
+        break;
+      case 'check':
+        if (!locator) break;
+        actionLines.push(`  await ${locator}.check();${comment}`);
+        break;
+      case 'uncheck':
+        if (!locator) break;
+        actionLines.push(`  await ${locator}.uncheck();${comment}`);
+        break;
+      case 'selectOption':
+        if (!locator) break;
+        actionLines.push(
+          `  await ${locator}.selectOption(${JSON.stringify(interaction.value ?? '')});${comment}`,
+        );
+        break;
+      default:
+        break;
+    }
+  }
+
+  let testLabel;
+  try {
+    const u = new URL(url);
+    testLabel = u.hostname + u.pathname;
+  } catch {
+    testLabel = url;
+  }
+
+  return [
+    `import { test, expect } from '@playwright/test';`,
+    ``,
+    `// Generated by copilot-devtools record-interactions`,
+    `// Source: ${url}`,
+    `// Captured: ${new Date().toISOString()}`,
+    ``,
+    `test(${JSON.stringify('recorded: ' + testLabel)}, async ({ page }) => {`,
+    `  await page.goto(${JSON.stringify(url)});`,
+    ...actionLines,
+    `});`,
+    ``,
+  ].join('\n');
+}
